+++ store-backend/productservice/src/main/java/com/praveenukkoji/productservice/service/ProductService.java
@@ -10,9 +10,9 @@ import com.praveenukkoji.productservice.exception.product.ProductCreateException
 import com.praveenukkoji.productservice.exception.product.ProductDeleteException;
 import com.praveenukkoji.productservice.exception.product.ProductNotFoundException;
 import com.praveenukkoji.productservice.exception.product.ProductUpdateException;
-import com.praveenukkoji.productservice.external.product.request.DecreaseProductStockRequest;
-import com.praveenukkoji.productservice.external.product.request.ProductDetailRequest;
-import com.praveenukkoji.productservice.external.product.response.ProductDetailResponse;
+import com.praveenukkoji.productservice.feign.product.request.DecreaseProductStockRequest;
+import com.praveenukkoji.productservice.feign.product.request.ProductDetailRequest;
+import com.praveenukkoji.productservice.feign.product.response.ProductDetailResponse;
 import com.praveenukkoji.productservice.model.Category;
 import com.praveenukkoji.productservice.model.Product;
 import com.praveenukkoji.productservice.repository.CategoryRepository;
